SURVEYS: UI: duplication cta for switching once created & user filter fix
